Move translation evaluation to view layer
Many important strings, like SubmissionResult.compilation_text,
Evaluation.text, UserTestResult.(compilation|evaluation)_text were not
translatable. They contained a lot of information (outcomes, time and
memory measures, stdout and stderr of the compiler, filenames, etc.)
which was formatted into them "deeply" inside CMS: by TaskTypes running
on the Worker, and then transmitted via Jobs to EvaluationService that
stored them in the database. Thus, fixing it required a lot of changes
throughout the codebase.

The compilation_text was split across many fields (_time,
_wall_clock_time, _memory, _stdout, _stderr), keeping only the outcome
in _text. All of the outcome fields now contain a JSON-encoded list
whose first element is a string with placeholders and whose other
elements (if any) are the data to fill in it. This allows the string to
be translated *by the frontend* before being formatted and shown.